Make separate integration tests, add a failing test and fix.
